New upstream version 5.2.0-dfsg

+++ src/VBox/Additions/linux/installer/vboxadd.sh
@@ -1,6 +1,6 @@
 #! /bin/sh
 #
-# Linux Additions kernel module init script ($Revision: 116785 $)
+# Linux Additions kernel module init script ($Revision: 118374 $)
 #
 
 #
@@ -51,11 +51,23 @@
 
 PATH=$PATH:/bin:/sbin:/usr/sbin
 PACKAGE=VBoxGuestAdditions
-LOG="/var/log/vboxadd-install.log"
 MODPROBE=/sbin/modprobe
 OLDMODULES="vboxguest vboxadd vboxsf vboxvfs vboxvideo"
-SCRIPTNAME=vboxadd.sh
+SERVICE="VirtualBox Guest Additions"
 QUICKSETUP=
+## systemd logs information about service status, otherwise do that ourselves.
+QUIET=
+
+setup_log()
+{
+    test -n "${LOG}" && return 0
+    # Rotate log files
+    LOG="/var/log/vboxadd-setup.log"
+    mv "${LOG}.3" "${LOG}.4" 2>/dev/null
+    mv "${LOG}.2" "${LOG}.3" 2>/dev/null
+    mv "${LOG}.1" "${LOG}.2" 2>/dev/null
+    mv "${LOG}" "${LOG}.1" 2>/dev/null
+}
 
 if $MODPROBE -c 2>/dev/null | grep -q '^allow_unsupported_modules  *0'; then
   MODPROBE="$MODPROBE --allow-unsupported-modules"
@@ -89,25 +101,30 @@ fi
 
 begin()
 {
-    test -n "${2}" && echo "${SCRIPTNAME}: ${1}."
-    logger -t "${SCRIPTNAME}" "${1}."
+    test -z "${QUIET}" && echo "${SERVICE}: ${1}"
 }
 
-succ_msg()
+info()
 {
-    logger -t "${SCRIPTNAME}" "${1}."
-}
-
-show_error()
-{
-    echo "${SCRIPTNAME}: failed: ${1}." >&2
-    logger -t "${SCRIPTNAME}" "${1}."
+    if test -z "${QUIET}"; then
+        echo "${SERVICE}: $1"
+    else
+        echo "$1"
+    fi
 }
 
 fail()
 {
-    show_error "$1"
+    log "${1}"
+    echo "$1" >&2
+    echo "The log file $LOG may contain further information." >&2
     exit 1
+}
+
+log()
+{
+    setup_log
+    echo "${1}" >> "${LOG}"
 }
 
 dev=/dev/vboxguest
@@ -194,7 +211,7 @@ do_vboxguest_non_udev()
 
 start()
 {
-    begin "Starting the VirtualBox Guest Additions" console;
+    begin "Starting."
     # If we got this far assume that the slow set-up has been done.
     QUICKSETUP=yes
     if test -z "${INSTALL_NO_MODULE_BUILDS}"; then
@@ -212,10 +229,8 @@ start()
 
             $MODPROBE vboxguest >/dev/null 2>&1 || {
                 setup
-                $MODPROBE vboxguest >/dev/null 2>&1 || {
-                    ${INSTALL_DIR}/init/vboxadd-x11 cleanup
+                $MODPROBE vboxguest >/dev/null 2>&1 ||
                     fail "modprobe vboxguest failed"
-                }
             }
             case "$no_udev" in 1)
                 sleep .5;;
@@ -228,17 +243,16 @@ start()
         running_vboxsf || {
             $MODPROBE vboxsf > /dev/null 2>&1 || {
                 if dmesg | grep "VbglR0SfConnect failed" > /dev/null 2>&1; then
-                    show_error "Unable to start shared folders support.  Make sure that your VirtualBox build"
-                    show_error "supports this feature."
+                    info "Unable to start shared folders support.  Make sure that your VirtualBox build supports this feature."
                 else
-                    show_error "modprobe vboxsf failed"
+                    info "modprobe vboxsf failed"
                 fi
             }
         }
     fi  # INSTALL_NO_MODULE_BUILDS
 
     # Put the X.Org driver in place.  This is harmless if it is not needed.
-    /sbin/rcvboxadd-x11 setup
+    "${INSTALL_DIR}/init/vboxadd-x11" setup 2>> "${LOG}"
     # Install the guest OpenGL drivers.  For now we don't support
     # multi-architecture installations
     rm -f /etc/ld.so.conf.d/00vboxvideo.conf
@@ -264,13 +278,12 @@ start()
     # This isn't necessary anymore as the vboxsf module is autoloaded.
     # mount -a -t vboxsf
 
-    succ_msg
     return 0
 }
 
 stop()
 {
-    begin "Stopping VirtualBox Additions" console;
+    begin "Stopping."
     if test -r /etc/ld.so.conf.d/00vboxvideo.conf; then
         rm /etc/ld.so.conf.d/00vboxvideo.conf
         ldconfig
@@ -280,11 +293,12 @@ stop()
     fi
     test -n "${INSTALL_NO_MODULE_BUILDS}" && return 0
     modprobe -q -r -a vboxvideo vboxsf vboxguest
-    egrep -q 'vboxguest|vboxsf|vboxvideo' /proc/modules &&
-        echo "You may need to restart your guest system to finish removing the guest drivers."
-    rm -f $userdev || fail "Cannot unlink $userdev"
-    rm -f $dev || fail "Cannot unlink $dev"
-    succ_msg
+    if egrep -q 'vboxguest|vboxsf|vboxvideo' /proc/modules; then
+        info "You may need to restart your guest system to finish removing the guest drivers."
+    else
+        rm -f $userdev || fail "Cannot unlink $userdev"
+        rm -f $dev || fail "Cannot unlink $dev"
+    fi
     return 0
 }
 
@@ -298,7 +312,7 @@ restart()
 # from the kernel as they may still be in use
 cleanup_modules()
 {
-    begin "Removing existing VirtualBox kernel modules"
+    log "Removing existing VirtualBox kernel modules."
     for i in ${OLDMODULES}; do
         # We no longer support DKMS, remove any leftovers.
         rm -rf "/var/lib/dkms/${i}"*
@@ -310,7 +324,6 @@ cleanup_modules()
         test -d "${i}" && rmdir -p "${i}" 2>/dev/null
     done
     rm -f /etc/depmod.d/vboxvideo-upstream.conf
-    succ_msg
 }
 
 # Build and install the VirtualBox guest kernel modules
@@ -320,34 +333,34 @@ setup_modules()
     test -z "${QUICKSETUP}" && cleanup_modules
     # This does not work for 2.4 series kernels.  How sad.
     test -n "${QUICKSETUP}" && test -f "${MODULE_DIR}/vboxguest.ko" && return 0
-    begin "Building the VirtualBox Guest Additions kernel modules"
+    info "Building the VirtualBox Guest Additions kernel modules."
 
-    begin "Building the main Guest Additions module"
+    # We are allowed to do ">> $LOG" after we have called "log()" once.
+    log "Building the main Guest Additions module."
     if ! $BUILDINTMP \
         --save-module-symvers /tmp/vboxguest-Module.symvers \
         --module-source $MODULE_SRC/vboxguest \
         --no-print-directory install >> $LOG 2>&1; then
-        show_error "Look at $LOG to find out what went wrong"
+        # If check_module_dependencies.sh fails it prints a message itself.
+        "${INSTALL_DIR}"/other/check_module_dependencies.sh 2>&1 &&
+            info "Look at $LOG to find out what went wrong"
         return 0
     fi
-    succ_msg
-    begin "Building the shared folder support module"
+    log "Building the shared folder support module"
     if ! $BUILDINTMP \
         --use-module-symvers /tmp/vboxguest-Module.symvers \
         --module-source $MODULE_SRC/vboxsf \
         --no-print-directory install >> $LOG 2>&1; then
-        show_error  "Look at $LOG to find out what went wrong"
+        info  "Look at $LOG to find out what went wrong"
         return 0
     fi
-    succ_msg
-    begin "Building the graphics driver module"
+    log "Building the graphics driver module"
     if ! $BUILDINTMP \
         --use-module-symvers /tmp/vboxguest-Module.symvers \
         --module-source $MODULE_SRC/vboxvideo \
         --no-print-directory install >> $LOG 2>&1; then
-        show_error "Look at $LOG to find out what went wrong"
+        info "Look at $LOG to find out what went wrong"
     fi
-    succ_msg
     [ -d /etc/depmod.d ] || mkdir /etc/depmod.d
     echo "override vboxguest * misc" > /etc/depmod.d/vboxvideo-upstream.conf
     echo "override vboxsf * misc" >> /etc/depmod.d/vboxvideo-upstream.conf
@@ -358,8 +371,7 @@ setup_modules()
 
 create_vbox_user()
 {
-    begin "Doing non-kernel setup of the Guest Additions"
-    echo "Creating user for the Guest Additions." >> $LOG
+    log "Creating user for the Guest Additions."
     # This is the LSB version of useradd and should work on recent
     # distributions
     useradd -d /var/run/vboxadd -g 1 -r -s /bin/false vboxadd >/dev/null 2>&1
@@ -372,7 +384,7 @@ create_udev_rule()
 {
     # Create udev description file
     if [ -d /etc/udev/rules.d ]; then
-        echo "Creating udev rule for the Guest Additions kernel module." >> $LOG
+        log "Creating udev rule for the Guest Additions kernel module."
         udev_call=""
         udev_app=`which udevadm 2> /dev/null`
         if [ $? -eq 0 ]; then
@@ -405,8 +417,7 @@ create_module_rebuild_script()
     cat << EOF > /etc/kernel/postinst.d/vboxadd
 #!/bin/sh
 test -d "/lib/modules/\${1}/build" || exit 0
-KERN_DIR="/lib/modules/\${1}/build" MODULE_DIR="/lib/modules/\${1}/misc" \
-/sbin/rcvboxadd quicksetup
+KERN_VER="\${1}" /sbin/rcvboxadd quicksetup
 exit 0
 EOF
     cat << EOF > /etc/kernel/prerm.d/vboxadd
@@ -428,17 +439,16 @@ shared_folder_setup()
     # Put the mount.vboxsf mount helper in the right place.
     ## @todo It would be nicer if the kernel module just parsed parameters
     # itself instead of needing a separate binary to do that.
-    ln -sf "$lib_path/$PACKAGE/mount.vboxsf" /sbin
+    ln -sf "${INSTALL_DIR}/other/mount.vboxsf" /sbin
     # SELinux security context for the mount helper.
     if test -e /etc/selinux/config; then
         # This is correct.  semanage maps this to the real path, and it aborts
         # with an error, telling you what you should have typed, if you specify
         # the real path.  The "chcon" is there as a back-up for old guests.
         command -v semanage > /dev/null &&
-            semanage fcontext -a -t mount_exec_t "/usr/lib/$PACKAGE/mount.vboxsf"
-        chcon -t mount_exec_t "$lib_path/$PACKAGE/mount.vboxsf"
+            semanage fcontext -a -t mount_exec_t "${INSTALL_DIR}/other/mount.vboxsf"
+        chcon -t mount_exec_t "${INSTALL_DIR}/other/mount.vboxsf"
     fi
-    succ_msg
 }
 
 # setup_script
@@ -447,7 +457,6 @@ setup()
     export BUILD_TYPE
     export USERNAME
 
-    rm -f $LOG
     MODULE_SRC="$INSTALL_DIR/src/vboxguest-$INSTALL_VER"
     BUILDINTMP="$MODULE_SRC/build_in_tmp"
     chcon -t bin_t "$BUILDINTMP" > /dev/null 2>&1
@@ -459,7 +468,7 @@ setup()
     test -n "${QUICKSETUP}" && return 0
     shared_folder_setup
     if  running_vboxguest || running_vboxadd; then
-        begin "Running kernel modules will not be replaced until the system is restarted"
+        info "Running kernel modules will not be replaced until the system is restarted"
     fi
     return 0
 }
@@ -479,7 +488,7 @@ cleanup()
     fi
 
     # Clean-up X11-related bits
-    ${INSTALL_DIR}/init/vboxadd-x11 cleanup
+    "${INSTALL_DIR}/init/vboxadd-x11" cleanup 2>> "${LOG}"
 
     # Remove other files
     rm /sbin/mount.vboxsf 2>/dev/null
@@ -499,6 +508,9 @@ dmnstatus()
     fi
 }
 
+case "$2" in quiet)
+    QUIET=yes;;
+esac
 case "$1" in
 start)
     start
@@ -524,7 +536,7 @@ status)
     dmnstatus
     ;;
 *)
-    echo "Usage: $0 {start|stop|restart|status|setup|quicksetup|cleanup}"
+    echo "Usage: $0 {start|stop|restart|status|setup|quicksetup|cleanup} [quiet]"
     exit 1
 esac
